modified the Update and delete user.
RequestValidator class: the email, name, and password methods public to be used to validate
Update request as well.

UserController class: modified the update method to take in a custom
request object instead of a User entity and modified the delete user
method to take in a imdbId instead of a user entity.

UserService class: Modified the update method to accept custom request
object, validate the fields and make sure that if a feild is present the
program doesn't crash or replace the data with the empty field. Modified
the delete method to take in a imdbId string instead of a user entitiy.